models/sdxl-turbo/requirements.txt
accelerate==1.6.0
certifi==2025.1.31
charset-normalizer==3.4.1
diffusers==0.33.1
filelock==3.18.0
fsspec==2025.3.2
grpcio==1.71.0
grpcio-tools==1.71.0
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pillow==11.2.1
protobuf==5.29.4
psutil==7.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==78.1.0
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.2
typing_extensions==4.13.2
urllib3==2.4.0
zipp==3.21.0
